# docs/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: HUAKAI Gateway API
  version: 0.1.0-pre
  summary: Multi-tenant relay-station AI gateway with provider account pooling.
  description: |
    HUAKAI exposes three client-protocol-compatible relay endpoints (OpenAI Chat
    Completions, OpenAI Responses, Anthropic Messages) plus an operator admin
    API for managing Pool Groups, Provider Accounts, Routes, Usage Records,
    Billing Ledger, and Audit trails.

    Source-of-truth specs at `docs/specs/`. This file is the Phase 2.2
    contract lock; Phase 3 implementation skeleton derives from this contract
    via codegen.

    Per [DR-001](../process/decisions/DR-001-multi-tenancy.md), every endpoint is
    tenant-scoped; tenant is server-resolved from API Key. Cross-tenant data
    access is structurally rejected.

    JSON convention: snake_case throughout (matches OpenAI / Anthropic public
    APIs and HUAKAI database schema).

    Money fields: numeric(20, 8) precision; rendered as JSON string for
    client-side BigNumber safety.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://huakai.example/api
    description: Operator-deployed instance (URL operator-configured)
security:
  - bearerAuth: []
...
paths:
...
  /v1/chat/completions:
    post:
      tags: [gateway]
      summary: OpenAI Chat Completions compatible relay
      operationId: postChatCompletions
      x-huakai-spec-source:
        - docs/specs/streaming-forwarder.md
        - docs/specs/protocol-translation.md
        - docs/specs/observability-billing.md
      description: |
        Accept OpenAI Chat Completions request, route to a Provider Account
        (F-POOL-001), translate via HUAKAI Canonical Stream Format (F-PROTO-002),
        return response (F-GW-002). Streaming via stream=true.

        Idempotency per F-OBS-001: when Idempotency-Key header is present, server
        derives idempotency_key from header + tenant_id + payload hash. Replays
        with same fingerprint return cached settled response.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IdempotentStreamReplay'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionsRequest'
      responses:
        '200':
          description: |
            Successful response. Content-Type is `application/json` for
            non-streaming or `text/event-stream` for streaming. SSE event
            types per F-PROTO-002 client adapter.
          headers:
            X-HUAKAI-Protocol-Loss:
              description: Comma-separated feature names that were LOSSY in this translation. Empty = fully PRESERVED. Full structured array also on Usage Record.
              schema: { type: string }
            X-HUAKAI-Idempotency-Hit:
              description: Present and 'true' when response is replayed from a settled prior claim with the same fingerprint.
              schema: { type: string, enum: ['true'] }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionsResponse'
            text/event-stream:
              schema:
                $ref: '#/components/schemas/SSEFrame'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceBusy' }
  /v1/responses:
    post:
      tags: [gateway]
      summary: OpenAI Responses compatible relay
      operationId: postResponses
      x-huakai-spec-source:
        - docs/specs/streaming-forwarder.md
        - docs/specs/protocol-translation.md
        - docs/specs/observability-billing.md
      description: |
        Same orchestration as `/v1/chat/completions` but client protocol is
        OpenAI Responses. Per F-PROTO-002 hub-and-spoke topology, both share
        the HUAKAI Canonical intermediate; only the client adapter differs.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IdempotentStreamReplay'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResponsesRequest' }
      responses:
        '200':
          description: Successful response (json or SSE per stream flag).
          headers:
            X-HUAKAI-Protocol-Loss: { schema: { type: string } }
            X-HUAKAI-Idempotency-Hit: { schema: { type: string, enum: ['true'] } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponsesResponse' }
            text/event-stream:
              schema: { $ref: '#/components/schemas/SSEFrame' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceBusy' }
  /v1/messages:
    post:
      tags: [gateway]
      summary: Anthropic Messages compatible relay
      operationId: postAnthropicMessages
      x-huakai-spec-source:
        - docs/specs/streaming-forwarder.md
        - docs/specs/protocol-translation.md
        - docs/specs/observability-billing.md
      description: |
        Anthropic Messages compatible client surface. Same hub-and-spoke
        orchestration as the OpenAI surfaces.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IdempotentStreamReplay'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnthropicMessagesRequest' }
      responses:
        '200':
          description: Successful response (json or SSE per stream flag).
          headers:
            X-HUAKAI-Protocol-Loss: { schema: { type: string } }
            X-HUAKAI-Idempotency-Hit: { schema: { type: string, enum: ['true'] } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnthropicMessagesResponse' }
            text/event-stream:
              schema: { $ref: '#/components/schemas/SSEFrame' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceBusy' }
...
  /v1/models:
    get:
      tags: [gateway]
      summary: OpenAI-compatible model discovery list
      operationId: listModels
      x-huakai-spec-source:
        - docs/process/plans/2026-04-30-n5-model-registry.md
      description: |
        Returns OpenAI-compatible model objects for public aliases that the
        authenticated tenant can route now: active alias, active model, and at
        least one enabled pool binding. Tenant-scoped disabled aliases block
        global inheritance just like chat request resolution.
      responses:
        '200':
          description: OpenAI-compatible model list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModelsListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceBusy' }
  /v1/me/usage:
    get:
      tags: [user-audit]
      summary: End-user self-service consumption records
      operationId: listMyUsage
      x-huakai-spec-source:
        - docs/specs/observability-billing.md
        - docs/specs/trust-chain-user-verifiable-ledger.md
      description: |
        Returns immutable usage records for the authenticated inbound API key.
        Tenant and API-key scope are resolved from bearer auth; client-supplied
        tenant_id, api_key_id, or user_id query parameters are ignored and do
        not affect the server-side scope. Response records intentionally omit
        tenant_id, api_key_id, user_id, prompt, body, and message payloads.
      parameters:
        - name: from
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: cursor
          in: query
          required: false
          description: Opaque cursor returned as `next_cursor`.
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
      responses:
        '200':
          description: Auth-scoped consumption records for the caller's API key.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeUsageListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceBusy' }
...
  /internal/runner/bootstrap:
    post:
      tags: [hermes-internal]
      summary: Issue an initial Hermes runner JWT
      operationId: bootstrapHermesRunnerJWT
      description: >-
        Internal endpoint that issues a short-lived EdDSA JWT for a Hermes runner.
        Requires HMAC proof: X-Hermes-Signature, X-Hermes-Timestamp, X-Hermes-Tenant,
        X-Hermes-User computed over `<timestamp>\n<method>\n<path>\n<raw_query>\n<tenant>\n<user>\n`
        + body, signed with HUAKAI_HERMES_INTERNAL_SHARED_SECRET HMAC-SHA256.
      security: []
      parameters:
        - name: X-Hermes-Signature
          in: header
          required: true
          schema: { type: string }
        - name: X-Hermes-Timestamp
          in: header
          required: true
          schema: { type: string }
        - name: X-Hermes-Tenant
          in: header
          required: true
          schema: { type: string }
        - name: X-Hermes-User
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [runner_id]
              properties:
                runner_id: { type: string, minLength: 1 }
                tenant_id: { type: integer, format: int64, minimum: 1 }
                actor_user_id: { type: integer, format: int64, minimum: 1 }
      responses:
        '200':
          description: Bootstrap JWT.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [token]
                properties:
                  token: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceBusy' }
  /internal/runner/refresh:
    post:
      tags: [hermes-internal]
      summary: Refresh a Hermes runner JWT
      operationId: refreshHermesRunnerJWT
      description: >-
        Internal endpoint that verifies an existing Hermes JWT and returns a replacement before expiry.
        Requires HMAC proof: X-Hermes-Signature, X-Hermes-Timestamp, X-Hermes-Tenant,
        X-Hermes-User computed over `<timestamp>\n<method>\n<path>\n<raw_query>\n<tenant>\n<user>\n`
        + body, signed with HUAKAI_HERMES_INTERNAL_SHARED_SECRET HMAC-SHA256.
      security: []
      parameters:
        - name: X-Hermes-Signature
          in: header
          required: true
          schema: { type: string }
        - name: X-Hermes-Timestamp
          in: header
          required: true
          schema: { type: string }
        - name: X-Hermes-Tenant
          in: header
          required: true
          schema: { type: string }
        - name: X-Hermes-User
          in: header
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Refreshed JWT.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [token]
                properties:
                  token: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceBusy' }
  /internal/keys:
    get:
      tags: [hermes-internal]
      summary: List active Hermes JWT public keys
      operationId: listHermesRunnerJWTKeys
      description: >-
        Internal public-key sync endpoint for Hermes runner local verifier cache.
        Requires HMAC proof: X-Hermes-Signature, X-Hermes-Timestamp, X-Hermes-Tenant,
        X-Hermes-User computed over `<timestamp>\n<method>\n<path>\n<raw_query>\n<tenant>\n<user>\n`
        + body, signed with HUAKAI_HERMES_INTERNAL_SHARED_SECRET HMAC-SHA256.
      security: []
      parameters:
        - name: X-Hermes-Signature
          in: header
          required: true
          schema: { type: string }
        - name: X-Hermes-Timestamp
          in: header
          required: true
          schema: { type: string }
        - name: X-Hermes-Tenant
          in: header
          required: true
          schema: { type: string }
        - name: X-Hermes-User
          in: header
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Active public keys.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [keys]
                properties:
                  keys:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      required: [kid, alg, public_key_pem]
                      properties:
                        kid: { type: string }
                        alg: { type: string, enum: [EdDSA] }
                        public_key_pem: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceBusy' }
...
  /v1/receipts/{request_id}:
    get:
      tags: [user-audit]
      summary: Fetch a tenant-scoped signed cost receipt
      operationId: getCostReceipt
      x-huakai-spec-source: [docs/specs/user-consumption-transparency.md]
      security:
        - sessionBearerAuth: []
      parameters:
        - in: path
          name: request_id
          required: true
          schema: { type: string, minLength: 1, maxLength: 256 }
      responses:
        '200':
          description: Signed user-verifiable receipt.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserCostReceipt' }
        '202':
          description: Receipt facts exist but final usage is still unavailable.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceBusy' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HUAKAI-API-KEY
    sessionBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HUAKAI-SESSION
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Client-supplied stable request id. Per docs/specs/observability-billing.md
        idempotency algorithm: server hashes (Idempotency-Key, tenant_id,
        api_key_id, payload, model, ...) into a settle-once claim. Replays
        with same fingerprint return the cached settled response.
      required: false
      schema: { type: string, maxLength: 128 }
...
  responses:
    BadRequest:
      description: Malformed or unsupported request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...